test(Refactoring): #2530 Updating import paths in spec files.
