+++ projects/igniteui-angular/src/lib/grids/grid/column-hiding.spec.ts
@@ -3,17 +3,18 @@ import { DebugElement } from '@angular/core';
 import { async, TestBed, fakeAsync, tick } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { IColumnVisibilityChangedEventArgs, IgxColumnHidingItemDirective } from './column-hiding-item.directive';
-import { IgxColumnHidingComponent, IgxColumnHidingModule } from './column-hiding.component';
-import { IgxGridModule, IgxGridComponent } from './index';
-import { IgxButtonModule } from '../directives/button/button.directive';
-import { ColumnDisplayOrder } from './column-chooser-base';
-import { ColumnHidingTestComponent, ColumnGroupsHidingTestComponent } from '../test-utils/grid-base-components.spec';
-import { UIInteractions, wait } from '../test-utils/ui-interactions.spec';
-import { GridFunctions } from '../test-utils/grid-functions.spec';
-import { HelperUtils } from '../test-utils/helper-utils.spec';
+import { IColumnVisibilityChangedEventArgs, IgxColumnHidingItemDirective } from '../column-hiding-item.directive';
+import { IgxColumnHidingComponent, IgxColumnHidingModule } from '../column-hiding.component';
+import { IgxGridModule } from '../index';
+import { IgxGridComponent } from './grid.component';
+import { IgxButtonModule } from '../../directives/button/button.directive';
+import { ColumnDisplayOrder } from '../column-chooser-base';
+import { ColumnHidingTestComponent, ColumnGroupsHidingTestComponent } from '../../test-utils/grid-base-components.spec';
+import { UIInteractions, wait } from '../../test-utils/ui-interactions.spec';
+import { GridFunctions } from '../../test-utils/grid-functions.spec';
+import { HelperUtils } from '../../test-utils/helper-utils.spec';
 
-import { configureTestSuite } from '../test-utils/configure-suite';
+import { configureTestSuite } from '../../test-utils/configure-suite';
 
 describe('Column Hiding UI', () => {
     configureTestSuite();
