+++ projects/igniteui-angular/src/lib/grids/grid/grid-filtering.spec.ts
@@ -2,13 +2,13 @@ import { Component, DebugElement, ViewChild } from '@angular/core';
 import { async, TestBed, fakeAsync, tick } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { Calendar, ICalendarDate } from '../calendar/calendar';
-import { FilteringLogic, IFilteringExpression } from '../data-operations/filtering-expression.interface';
+import { Calendar, ICalendarDate } from '../../calendar/calendar';
+import { FilteringLogic, IFilteringExpression } from '../../data-operations/filtering-expression.interface';
 import { IgxGridComponent } from './grid.component';
-import { IgxGridModule } from './index';
+import { IgxGridModule } from '../index';
 import { IgxStringFilteringOperand, IgxNumberFilteringOperand,
-    IgxBooleanFilteringOperand, IgxDateFilteringOperand, IgxFilteringOperand, FilteringExpressionsTree } from '../../public_api';
-import { configureTestSuite } from '../test-utils/configure-suite';
+    IgxBooleanFilteringOperand, IgxDateFilteringOperand, IgxFilteringOperand, FilteringExpressionsTree } from '../../../public_api';
+import { configureTestSuite } from '../../test-utils/configure-suite';
 
 const FILTERING_TOGGLE_CLASS = 'igx-filtering__toggle';
 const FILTERING_TOGGLE_FILTERED_CLASS = 'igx-filtering__toggle--filtered';
