+++ projects/igniteui-angular/src/lib/grids/grid/grid-summary.spec.ts
@@ -2,13 +2,14 @@
 import { async, fakeAsync, TestBed, tick } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { BrowserAnimationsModule, NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { IgxInputDirective } from '../directives/input/input.directive';
-import { IgxDateSummaryOperand, IgxGridComponent, IgxGridModule, IgxNumberSummaryOperand, IgxSummaryResult } from './index';
-import { IgxGridAPIService } from './api.service';
-import { UIInteractions, wait } from '../test-utils/ui-interactions.spec';
-import { GridFunctions } from '../test-utils/grid-functions.spec';
+import { IgxInputDirective } from '../../directives/input/input.directive';
+import { IgxDateSummaryOperand, IgxGridModule, IgxNumberSummaryOperand, IgxSummaryResult } from '../index';
+import { IgxGridComponent } from './grid.component';
+import { IgxGridAPIService } from './grid-api.service';
+import { UIInteractions, wait } from '../../test-utils/ui-interactions.spec';
+import { GridFunctions } from '../../test-utils/grid-functions.spec';
 
-import { configureTestSuite } from '../test-utils/configure-suite';
+import { configureTestSuite } from '../../test-utils/configure-suite';
 
 describe('IgxGrid - Summaries', () => {
     configureTestSuite();
