+++ projects/igniteui-angular/src/lib/grids/grid/grid.search.spec.ts
@@ -1,15 +1,15 @@
 import { async, TestBed, fakeAsync, tick } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { IgxGridModule } from './index';
-import { SortingDirection } from '../data-operations/sorting-expression.interface';
-import { IgxStringFilteringOperand } from '../../public_api';
-import { BasicGridSearchComponent } from '../test-utils/grid-base-components.spec';
-import { SampleTestData } from '../test-utils/sample-test-data.spec';
-import { GridWithAvatarComponent, GroupableGridSearchComponent, ScrollableGridSearchComponent } from '../test-utils/grid-samples.spec';
-import { IForOfState } from '../directives/for-of/for_of.directive';
-import { wait } from '../test-utils/ui-interactions.spec';
+import { IgxGridModule } from '../index';
+import { SortingDirection } from '../../data-operations/sorting-expression.interface';
+import { IgxStringFilteringOperand } from '../../../public_api';
+import { BasicGridSearchComponent } from '../../test-utils/grid-base-components.spec';
+import { SampleTestData } from '../../test-utils/sample-test-data.spec';
+import { GridWithAvatarComponent, GroupableGridSearchComponent, ScrollableGridSearchComponent } from '../../test-utils/grid-samples.spec';
+import { IForOfState } from '../../directives/for-of/for_of.directive';
+import { wait } from '../../test-utils/ui-interactions.spec';
 
-import { configureTestSuite } from '../test-utils/configure-suite';
+import { configureTestSuite } from '../../test-utils/configure-suite';
 
 describe('IgxGrid - search API', () => {
     configureTestSuite();
